Add files via upload
First, I created an HTML file which is containing the To Do List that I created with Bootstrap and added these files here as well. 
Then I made this HTML file interactive using the DOM. It is possible to add and remove the To-Do List that I have created. There is also button to delete all items.